+++ src/pages/soobinLee/Main/FOOTER_INFO.js
@@ -1,0 +1,37 @@
+export const FOOTER_INFO_LIST = [
+  {
+    id: 1,
+    href: 'http://localhost:3001/MainSoobin',
+    title: 'Instagram 정보',
+  },
+
+  {
+    id: 2,
+    href: 'http://localhost:3001/MainSoobin',
+    title: '∙ 지원 ∙ 홍보 센터',
+  },
+
+  {
+    id: 3,
+    href: 'http://localhost:3001/MainSoobin',
+    title: '∙ API ∙ 채용 정보 ',
+  },
+
+  {
+    id: 4,
+    href: 'http://localhost:3001/MainSoobin',
+    title: '∙ 개인정보처리방침 ∙ 약관',
+  },
+
+  {
+    id: 5,
+    href: 'http://localhost:3001/MainSoobin',
+    title: '∙ 디렉터리 ∙ 프로필 ∙ 해시태그',
+  },
+
+  {
+    id: 6,
+    href: 'http://localhost:3001/MainSoobin',
+    title: '∙ 언어 © 2023 INSTAGRAM',
+  },
+];
