+++ src/pages/soobinLee/Main/MainSoobin.scss
@@ -302,6 +302,18 @@ img.comment_heart {
   margin-right: 14px;
 }
 
+.comment_like {
+  display: flex;
+}
+
+img.delete_icon {
+  display: flex;
+  position: relative;
+  width: 22px;
+  height: 22px;
+  margin-top: -6px;
+}
+
 /*input comment box*/
 
 .input_box {
@@ -339,6 +351,12 @@ button.submit-comment {
   color: #b9dffc;
 }
 
+.deleteBtn {
+  border-style: none;
+  background-color: #fff;
+  margin-right: 3px;
+}
+
 /*main right*/
 
 .main_right {
@@ -346,6 +364,9 @@ button.submit-comment {
   right: 220px;
   width: 380px;
   height: 450px;
+  position: sticky;
+  top: 0;
+  right: 0;
 }
 
 .useprofilebox {
@@ -513,7 +534,15 @@ button.submit-comment {
 }
 
 .script {
-  margin-top: 20px;
-  font-size: 16px;
+  display: inline-block;
+  margin-right: 1px;
+  text-align: center;
+  padding: 7px;
+}
+
+.script2 {
+  display: flex;
+  justify-content: space-between;
+  text-decoration-line: none;
   color: #adb5bd;
 }
